backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=1.0.0
langchain-core>=1.0.0
langchain-text-splitters>=1.0.0
langchain-chroma>=0.1.0
langgraph>=0.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
wikipedia-api>=0.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0